Improve clarity of error messages
If there the line is 0 or falsy, we print only the message now. As those
sorts of messages don't really have an associated position, they're
about the whole document.

We're also saying "on or around line" instead of "on line" to indicate
we don't know it's exactly this line yet.