Change example field order (#54)
While the current field order represents the order displayed in Biz Ops, its the troubleshooting information which is most useful in a runbook, so should be given more prominance than things like GDPR/personal data. It's also likely that high quality information will be filled in in earlier fields, if entering a large body of data I find the quality and quanitity of my writing tails off by the end.